New Activity Added
WallpaperSetActivity Added
Improve UI
Add Permission
