+++ android/app/src/main/java/ir/amirsobhan/wallpaperapp/Adapter/WallpaperAdapter.java
@@ -2,7 +2,6 @@ package ir.amirsobhan.wallpaperapp.Adapter;
 
 import android.content.Context;
 import android.content.Intent;
-import android.content.res.Configuration;
 import android.graphics.drawable.AnimatedVectorDrawable;
 import android.os.Build;
 import android.util.Log;
@@ -11,7 +10,6 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
 import android.widget.TextView;
-import android.widget.Toast;
 
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
@@ -34,7 +32,7 @@ import java.util.List;
 import java.util.Map;
 
 import ir.amirsobhan.wallpaperapp.Databases.WallpaperDB;
-import ir.amirsobhan.wallpaperapp.FullscreenViewActivity;
+import ir.amirsobhan.wallpaperapp.WallpaperSetActivity;
 import ir.amirsobhan.wallpaperapp.Model.ApiResult;
 import ir.amirsobhan.wallpaperapp.Model.Wallpaper;
 import ir.amirsobhan.wallpaperapp.R;
@@ -121,7 +119,7 @@ public class WallpaperAdapter extends RecyclerView.Adapter<WallpaperAdapter.Wall
         holder.itemView.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                Intent intent = new Intent(context,FullscreenViewActivity.class);
+                Intent intent = new Intent(context, WallpaperSetActivity.class);
                 intent.putExtra("json",new Gson().toJson(wallpaper));
                 context.startActivity(intent);
             }
