+++ android/app/src/main/res/menu/full_menu.xml
@@ -1,9 +1,17 @@
 <?xml version="1.0" encoding="utf-8"?>
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto">
+
+    <item
+        android:id="@+id/set_wallpaper"
+        android:title="Set Wallpaper"
+        app:showAsAction="always"
+        android:icon="@drawable/ic_baseline_save_24"/>
+
     <item
         android:id="@+id/go_fullscreen"
         android:title="FulllScreen"
         app:showAsAction="always"
         android:icon="@drawable/ic_baseline_fullscreen_24"/>
+
 </menu>
